+++ src/containers/BooksList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 const BooksList = () => {
   const { books } = useSelector((state) => state);
@@ -16,7 +16,7 @@ const BooksList = () => {
         </tr>
       </thead>
       <tbody>
-        {books.books.map((book) => (
+        {books.map((book) => (
           <Book key={book.id} title={book.title} category={book.category} />
         ))}
       </tbody>
