--- src/containers/store.js
@@ -1,6 +1,0 @@
-import { createStore } from 'redux';
-import reducer from './reducers';
-
-const store = createStore(reducer);
-
-export default store;
